update file structure, make page width dynamic
