+++ packages/pdf-viewer/pages.css
@@ -1,0 +1,76 @@
+::-webkit-scrollbar {
+    width: 7px;
+    height: 7px;
+}
+
+::-webkit-scrollbar-corner {
+    background: none;
+}
+
+::-webkit-scrollbar-track {
+    border: none;
+}
+
+::-webkit-scrollbar-thumb {
+    background: var(--grey);
+    border-radius: 5px;
+}
+
+.app-pages {
+    display: flex;
+    justify-content: center;
+    align-items: start;
+    overflow-x: hidden;
+    width: 100%;
+    height: 100%;
+    padding: 0.5rem;
+    padding-top: 0px;
+    overflow-y: hidden;
+    padding-right: 0.25rem;
+    position: relative;
+}
+
+.app-pages.focused {
+    padding-right: 0px;
+    overflow-y: auto;
+}
+
+.pages-holder {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    overflow: hidden;
+    flex-flow: column;
+    width: 100%;
+    min-height: 100%;
+    height: auto;
+    row-gap: 2px;
+}
+
+.page-wrapper {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    overflow: hidden;
+    border: solid thin rgba(120, 120, 120, 0.498);
+    background: rgb(233, 233, 233);
+    position: relative;
+}
+
+.p-info {
+    position: absolute;
+    top: 0.5rem;
+    left: 0.5rem;
+    padding: 0.3rem;
+    background: rgba(203, 203, 203, 0.509);
+    border-radius: 0.3rem;
+    font-size: 0.8rem;
+    color: rgba(91, 91, 91, 0.829);
+    backdrop-filter: blur(0.5rem);
+    cursor: default;
+}
+
+.p-info:hover {
+    opacity: 0.3;
+}
