+++ packages/pdf-viewer/pdfSource.ts
@@ -1,0 +1,70 @@
+const pdfjsLib = require('pdfjs-dist');
+
+
+export interface ScaledSize {
+	height: number;
+	width: number;
+	scale: number;
+}
+
+export class PdfData {
+	public url: string;
+	private doc: any = null;
+	public pageCount: number = null;
+	private pages: any = {};
+	private pageSize: {
+		height: number;
+		width: number;
+	} = null;
+	public error: any = null;
+	public constructor() {
+	}
+	public loadDoc = async (url: string) => {
+		this.url = url;
+		const loadingTask = pdfjsLib.getDocument(url);
+		try {
+			const pdfDocument: any = await loadingTask.promise;
+			this.doc = pdfDocument;
+			this.pageCount = pdfDocument.numPages;
+		} catch (err) {
+			console.error(`Error: ${err}`);
+			this.error = err;
+		}
+	};
+	public getPage = async (pageNo: number) => {
+		if (!this.doc) {
+			throw new Error('Document not loaded');
+		}
+		if (!this.pages[pageNo]) {
+			this.pages[pageNo] = await this.doc.getPage(pageNo);
+		}
+		return this.pages[pageNo];
+	};
+	public getPageSize = async () => {
+		if (!this.pageSize) {
+			const page = await this.getPage(1);
+			const viewport = page.getViewport({ scale: 1.0 });
+			this.pageSize = {
+				height: viewport.height,
+				width: viewport.width,
+			};
+		}
+		return this.pageSize;
+	};
+	public getScaledSize = async (height: number = null, width: number = null): Promise<ScaledSize> => {
+		const actualSize = await this.getPageSize();
+		let scale = 1.0;
+		if (height && width) {
+			scale = Math.min(height / actualSize.height, width / actualSize.width);
+		} else if (height) {
+			scale = height / actualSize.height;
+		} else if (width) {
+			scale = width / actualSize.width;
+		}
+		return {
+			height: actualSize.height * scale,
+			width: actualSize.width * scale,
+			scale,
+		};
+	};
+}
